remove MCSAaccessMode and split swarm.go and use NodeRole
